Incorporated review comments.
